Created BrowserWindowFeatures subclass
brave/browser/ui/browser_window/sources.gni is used because
having //brave/browser/ui/browser_window target causes dependency
cycle with //chrome/browser/ui/browser_window.
