+++ browser/ui/browser_window/browser_window_features.cc
@@ -1,0 +1,64 @@
+/* Copyright (c) 2025 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#include "chrome/browser/ui/browser_window/public/browser_window_features.h"
+
+#include "base/check_is_test.h"
+#include "base/memory/ptr_util.h"
+#include "base/no_destructor.h"
+#include "brave/components/brave_vpn/common/buildflags/buildflags.h"
+
+#if BUILDFLAG(ENABLE_BRAVE_VPN)
+#include "brave/browser/ui/brave_vpn/brave_vpn_controller.h"
+#endif
+
+namespace {
+
+// This is the generic entry point for test code to stub out browser window
+// functionality. It is called by production code, but only used by tests.
+BrowserWindowFeatures::BrowserWindowFeaturesFactory& GetFactory() {
+  static base::NoDestructor<BrowserWindowFeatures::BrowserWindowFeaturesFactory>
+      factory;
+  return *factory;
+}
+
+}  // namespace
+
+// static
+std::unique_ptr<BrowserWindowFeatures>
+BrowserWindowFeatures::CreateBrowserWindowFeatures() {
+  if (GetFactory()) {
+    CHECK_IS_TEST();
+    return GetFactory().Run();
+  }
+  // Constructor is protected.
+  return base::WrapUnique(new BrowserWindowFeatures());
+}
+
+// static
+void BrowserWindowFeatures::ReplaceBrowserWindowFeaturesForTesting(
+    BrowserWindowFeaturesFactory factory) {
+  BrowserWindowFeatures::BrowserWindowFeaturesFactory& f = GetFactory();
+  f = std::move(factory);
+}
+
+BrowserWindowFeatures::BrowserWindowFeatures() = default;
+BrowserWindowFeatures::~BrowserWindowFeatures() = default;
+
+void BrowserWindowFeatures::InitPostBrowserViewConstruction(
+    BrowserView* browser_view) {
+  BrowserWindowFeatures_ChromiumImpl::InitPostBrowserViewConstruction(
+      browser_view);
+
+#if BUILDFLAG(ENABLE_BRAVE_VPN)
+  brave_vpn_controller_ = std::make_unique<BraveVPNController>(browser_view);
+#endif
+}
+
+#if BUILDFLAG(ENABLE_BRAVE_VPN)
+BraveVPNController* BrowserWindowFeatures::GetBraveVPNController() {
+  return brave_vpn_controller_.get();
+}
+#endif
