+++ browser/ui/browser_window/sources.gni
@@ -1,0 +1,10 @@
+# Copyright (c) 2025 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+brave_browser_ui_browser_window_sources =
+    [ "//brave/browser/ui/browser_window/public/browser_window_features.h" ]
+
+brave_browser_ui_browser_window_deps =
+    [ "//brave/components/brave_vpn/common/buildflags" ]
